Suspendable Spark Tasks using Thread wait/notify (#26)
* Adding ThreadSync |Functionality

* SparkConf - make sure one of Coroutines/ThreadSync is enabled
Adding test case to SparkConf

* *  Extending Task and DAG functionality to support both coroutine and threadSync primitives.
*  Currently passing two Boolean values to distinguish between them - could be improved by using an enum.

* ThreadSync implementation of RDD api + TestSuite

* ThreadSync tests for DataFrame and Datasets

* Adding ThreadSync SparkConf - make sure one of Coroutines/ThreadSync is enabled
Adding test case to SparkConf

* *  Extending Task and DAG functionality to support both coroutine and threadSync primitives.
*  Currently passing two Boolean values to distinguish between them - could be improved by using an enum.

* ThreadSync implementation of RDD api + TestSuite

* ThreadSync tests for DataFrame and Datasets

* Basic operator Implementation with wait/notify
Adding RDD Test suite

* DataSet + DataFrame testing

* Adding needed functionality for Suspendable Tasks in the Executor and TaskContext

* Different shuffle threadSync implementations - SparkContextSuite testing new functionality

* Style FIX, interrupt thread when pause - false by default

* Reusing ThreadSync wait mechanism across operators

* Pause/Resume executor events now track latency (measured in nanosends) and propagate the appropriate events back to the driver in separate threads.

* FIX ensure proper yield/resume variable Type on Executor

* Tracking Task Pause/Resume latency (measured in nanosends) in the TaskRunner and propagate the appropriate events back to the driver.
Added (task-runner) thread-safe methods for updating metrics and Inline Threads that track the progress for ThreadSync case.

* TaskCntx IsPaused method needs to accessible for submitting custom Coroutine functions

* FIX: double offered resources check in CoarseGrainedScheduler banckend - extra manual-scheduling check
Refactor: Executor Pause/Resume method logic moved to TaskRunner - better logging

* ThreadSync: Moving setTaskResumedEndTime outside isPaused check in checkSuspend.
Executor: removing unreachable condition

* FIX KafkaWriter ThreadSync

* Make sure there is a single Pause/Resume notifier per task (issue caused when we pause/resume a task multiple times)

* Extra scheduling policy test case to reproduce bug

* FIX LB policy now only suspends tasks when there are not enough free cores available

* * Merge two Cache-aware scheduling policies to 1
* FIX tests that used old pause/resume Double values
* Adding new policy tests being aware of free resources on Nodes + LB

+++ core/src/main/scala/org/apache/spark/scheduler/TaskInfo.scala
@@ -69,11 +69,11 @@ class TaskInfo(
    * The time when the task has completed successfully (including the time to remotely fetch
    * results, if necessary).
    */
-  var finishTime: Long = 0
-
-  var pauseLatency: Double = 0.0
-
-  var resumeLatency: Double = 0.0
+  var finishTime: Long = 0L
+  // In nanoseconds
+  var pauseLatency: Long = 0L
+  // In nanoseconds
+  var resumeLatency: Long = 0L
 
   var failed = false
